+++ customk/classes/admin.py
@@ -1,14 +1,16 @@
+from typing import Optional, Dict, Any
+from django.http import HttpRequest, HttpResponse
 from django.contrib import admin, messages
 
 from .models import Class, ClassDate, ClassImages, ExchangeRate
 
 
-class ClassDateInline(admin.TabularInline):
+class ClassDateInline(admin.TabularInline): # type: ignore
     model = ClassDate
     extra = 0
 
 
-class ClassImagesInline(admin.TabularInline):
+class ClassImagesInline(admin.TabularInline): # type: ignore
     model = ClassImages
     extra = 0
 
@@ -27,21 +29,21 @@ class ClassAdmin(admin.ModelAdmin):  # type: ignore
         "average_rating",
     )
 
-    def price_in_usd(self, obj):
+    def price_in_usd(self, obj: Class) -> Optional[float]:
         usd_price = obj.get_price_in_usd()
         return usd_price
 
-    price_in_usd.short_description = "가격 (USD)"
+    price_in_usd.short_description = "가격 (USD)" # type: ignore
 
-    def is_viewed_badge(self, obj):
+    def is_viewed_badge(self, obj: Class) -> str:
         if obj.is_viewed:
             return "✅"
         else:
             return "❗"
 
-    is_viewed_badge.short_description = "조회 상태"
+    is_viewed_badge.short_description = "조회 상태" # type: ignore
 
-    def changelist_view(self, request, extra_context=None):
+    def changelist_view(self, request: HttpRequest, extra_context: Optional[Dict[str, Any]]=None) -> HttpResponse:
         unviewed_classes_count = Class.objects.filter(is_viewed=False).count()
 
         if unviewed_classes_count > 0:
@@ -52,7 +54,7 @@ class ClassAdmin(admin.ModelAdmin):  # type: ignore
 
         return super().changelist_view(request, extra_context=extra_context)
 
-    def change_view(self, request, object_id, form_url="", extra_context=None):
+    def change_view(self, request: HttpRequest, object_id: str, form_url: str="", extra_context: Optional[Dict[str, Any]] = None) -> HttpResponse:
         course = self.get_object(request, object_id)
         if course and not course.is_viewed:
             course.is_viewed = True
@@ -64,5 +66,5 @@ class ClassAdmin(admin.ModelAdmin):  # type: ignore
 
 
 @admin.register(ExchangeRate)
-class ExchangeRateAdmin(admin.ModelAdmin):
+class ExchangeRateAdmin(admin.ModelAdmin): # type: ignore
     list_display = ("currency", "rate")
