+++ customk/classes/models.py
@@ -1,6 +1,7 @@
 from django.core.serializers.json import DjangoJSONEncoder
 from django.db import models
 from django.db.models import Avg
+from typing import Optional, Dict, Any
 
 from common.models import CommonModel
 
@@ -9,7 +10,7 @@ class ExchangeRate(models.Model):
     currency = models.CharField(max_length=10, default="USD")
     rate = models.DecimalField(max_digits=10, decimal_places=4)
 
-    def __str__(self):
+    def __str__(self) -> str:
         return f"{self.currency}: {self.rate}"
 
 
@@ -23,22 +24,25 @@ class Class(CommonModel):
 
     is_viewed = models.BooleanField(default=False)
 
-    def __str__(self):
+    def __str__(self)-> str:
         return self.title
 
-    def get_price_in_usd(self):
-        exchange_rate = ExchangeRate.objects.filter(currency="USD").first()
+    def get_price_in_usd(self)-> Optional[float]:
+        exchange_rate: Optional[ExchangeRate] = ExchangeRate.objects.filter(currency="USD").first()
         if exchange_rate:
-            return round(self.price / exchange_rate.rate, 2)
+            return round(float(self.price) / float(exchange_rate.rate), 2)
         return None
 
     @property
-    def average_rating(self):
-        avg = self.reviews.aggregate(average=Avg("rating"))["average"] or 0
-        return round(avg, 1)
+    def average_rating(self)-> Optional[float]:
+        avg: Optional[float] = self.reviews.aggregate(average=Avg("rating"))["average"] or 0
+
+        if avg is None:
+            return None
+        return round(float(avg), 1)
 
     @property
-    def formatted_address(self):
+    def formatted_address(self) -> str:
         address = self.address or {}
         state = address.get("state", "")
         city = address.get("city", "")
@@ -58,5 +62,5 @@ class ClassImages(models.Model):
     class_id = models.ForeignKey(Class, related_name="images", on_delete=models.CASCADE)
     image_url = models.URLField(max_length=2000)
 
-    def __str__(self):
+    def __str__(self) -> str:
         return f"{self.class_id.title}"
