+++ customk/questions/urls.py
@@ -1,6 +1,7 @@
 from django.urls import path
 
-from .views import AnswerListView, QuestionListView
+from questions.views.answer import AnswerListView
+from questions.views.question import QuestionListView
 
 urlpatterns = [
     path("<int:class_id>/", QuestionListView.as_view(), name="class-question-list"),
