+++ customk/reactions/models.py
@@ -1,7 +1,8 @@
 from django.db import models
 from django.db.models import Count, Q
-
+from typing import Dict
 from common.models import CommonModel
+from reviews.models import Review
 
 
 class Reaction(CommonModel):
@@ -21,7 +22,7 @@ class Reaction(CommonModel):
     reaction = models.IntegerField(choices=REACTON_CHOICES, default=NO_REACTION)
 
     @staticmethod
-    def get_review_reactions(review):
+    def get_review_reactions(review: Review) -> Dict[str, int]:
         reactions = Reaction.objects.filter(review=review).aggregate(
             likes_count=Count("pk", filter=Q(reaction=Reaction.LIKE)),
             dislikes_count=Count("pk", filter=Q(reaction=Reaction.DISLIKE)),
